+++ ichnaea/api/locate/locate_v2/tests.py
@@ -103,10 +103,9 @@ class TestView(LocateV2Base, CommonLocateTest, CommonPositionTest):
             (self.metric_type + '.request', [self.metric_path, 'key:test']),
             (self.metric_type + '.result',
                 ['key:test', 'country:none', 'accuracy:medium', 'status:hit']),
-            # TODO
-            # (self.metric_type + '.source',
-            #     ['key:test', 'country:none', 'source:internal',
-            #      'accuracy:medium', 'status:hit']),
+            (self.metric_type + '.source',
+                ['key:test', 'country:none', 'source:internal',
+                 'accuracy:medium', 'status:hit']),
         ], timer=[
             ('request', [self.metric_path, 'method:post']),
         ])
@@ -153,10 +152,9 @@ class TestView(LocateV2Base, CommonLocateTest, CommonPositionTest):
             (self.metric_type + '.request', [self.metric_path, 'key:test']),
             (self.metric_type + '.result',
                 ['key:test', 'country:none', 'accuracy:high', 'status:hit']),
-            # TODO
-            # (self.metric_type + '.source',
-            #     ['key:test', 'country:none', 'source:internal',
-            #      'accuracy:high', 'status:hit']),
+            (self.metric_type + '.source',
+                ['key:test', 'country:none', 'source:internal',
+                 'accuracy:high', 'status:hit']),
         ], timer=[
             ('request', [self.metric_path, 'method:post']),
         ])